requirements.txt
streamlit==1.24.1
torch==2.0.1
transformers==4.30.2
Pillow==10.0.0
numpy==1.25.2
